Create StellarMotion.mat
The spectra data was collected at evenly-spaced wavelengths, and you know the starting wavelength (λ
start), the spacing (λ delta), and the number of observations. Each column of spectra is the spectrum of a different star. The second column is the spectrum of star HD 94028. Use the loglog function in the same way as the plot function to plot data using a log scales for each axis.loglog(x,y,"*--"). Recall that the min function allows two outputs, the second of which is the index at which the minimum value occurred. This index corresponds to the location of the Hydrogen-alpha line. The line (lambdaHa,sHa) is the location of the Hydrogen-alpha line. If you zoom in on the plot, you can see that the wavelength of the Hydrogen-alpha line of HD 94028 is 656.62 nm, which is slightly longer than the laboratory value of 656.28 nm.Using the Hydrogen-alpha wavelength of the star, you can calculate the redshift factor (the speed of the star relative to the earth) using the formula z=(λHa/656.28)−1. You can then calculate the speed by multiplying the redshift factor by the speed of light (299792.458km/s). Having created a live script for finding the redshift, you can easily modify the script to repeat the calculation on any star in the spectra matrix.