Create StellarMotion2.mat
On line 2 of the displayed script, the spectrum data is extracted for the second star in the matrix spectra. Then lines 3 through 5 calculate the speed based on that data. To calculate the speed of all the stars in spectra, You could repeat the calculations in a for loop, but it is more efficient to use array operations instead. Notice that speed is now a vector. A positive speed means that the star is moving away from us (redshifted spectrum), and a negative speed means the star is moving towards us (blueshifted spectrum).
In the next step I created a plot containing all seven stars. I'll use different styles for redshifted and blueshifted spectra. Since the plot command won't be the same for every star, it's convenient to use a for loop. First, I'll plot the blueshifted spectra using dashed lines. I'll plot the redshifted spectra using a thick line. You can pass a string array directly to the legend function. The string array starnames contains the name of each star in spectra. In the plot, you can use the line styles to identify the stars with redshifted spectra, then look up their names in the legend. you determine the names of the redshifted spectra without a for loop for that, Recall that you can use logical indexing to find elements that match a condition. c = b(a < 6)